reorganize featurizer module
