+++ chemprop/v2/featurizers/mixins.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import Mapping, Optional, Sequence, Union
+from typing import Optional, Sequence, Union
 
 import numpy as np
 from rdkit import Chem
@@ -26,6 +26,12 @@ class MolGraphFeaturizerMixin(ABC):
     def __call__(self, *args, **kwargs):
         return self.featurize(*args, **kwargs)
 
+    @property
+    def shape(self) -> tuple[int, int]:
+        """the feature dimension of the atoms and bonds, respectively, of `MolGraph`s generated by
+        this featurizer"""
+        return self.atom_fdim, self.bond_fdim
+
     @abstractmethod
     def featurize(
         self,
@@ -34,3 +40,37 @@ class MolGraphFeaturizerMixin(ABC):
         bond_features_extra: Optional[np.ndarray] = None,
     ) -> MolGraph:
         pass
+
+
+class MultiHotFeaturizerMixin(ABC):
+    """A `MultiHotFeaturizer` calculates feature vectors of arbitrary objects by concatenating
+    multiple one-hot feature vectors"""
+
+    def __call__(self, x) -> np.ndarray:
+        return self.featurize(x)
+
+    @abstractmethod
+    def __len__(self) -> int:
+        """the length of a feature vector from this featurizer"""
+
+    @property
+    @abstractmethod
+    def subfeatures(self) -> Sequence[tuple[str, slice]]:
+        """a list of tuples containing the subfeature name and its respective slice in the output 
+        feature vectors"""
+
+    @abstractmethod
+    def featurize(self, x) -> np.ndarray:
+        """calculate the feature vector of x"""
+
+    @property
+    def num_subfeatures(self) -> int:
+        return len(self.subfeatures)
+
+    @staticmethod
+    def one_hot_index(x, xs: Sequence) -> tuple[int, int]:
+        """return the index of a one hot encoding of `x` given choices `xs` and the length of the
+        uncompressed encoding"""
+        n = len(xs)
+        
+        return xs.index(x) if x in xs else n, n + 1
